add channels + possibility to join them + send/recv messages

+++ includes/core/sockets/SockStream.hpp
@@ -56,7 +56,7 @@ class SockStream
 		void						close(void);
 		void						setPackageProtocol(IProtocol &proto);
 
-		int							getSocket(void) const;
+		ushort						getSocket(void) const;
 		const struct sockaddr_in&	getAddress(void) const;
 		IProtocol					*getProtocol(void) const;
 		Package						&getRecievedData(void);
